+++ src/com/rentarosato520/spacefighter/entity/ships/StealthDrone.java
@@ -8,7 +8,7 @@ import com.rentarosato520.spacefighter.entity.EntityObject;
 
 public class StealthDrone extends EntityObject{
 
-	public StealthDrone(int x, int y, int width, int height, Handler h) {
+	public StealthDrone(float x, float y, int width, int height, Handler h) {
 		super(x, y, width, height, h);
 	}
 
@@ -23,7 +23,7 @@ public class StealthDrone extends EntityObject{
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, width, height);
+		return new Rectangle((int)x,(int) y, width, height);
 	}
 
 }
